web_chat/requirements.txt
fastapi
uvicorn
jinja2
groq
pydantic
python-multipart
psycopg2-binary